♻️: Move utility classes to selidor-core
