+++ .idea/encodings.xml
@@ -5,7 +5,10 @@
     <file url="file://$PROJECT_DIR$/selidor-projects" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/selidor-projects/selidor-dependencies" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/selidor-projects/selidor-libraries" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/selidor-projects/selidor-libraries/selidor-core" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/selidor-projects/selidor-libraries/selidor-identifier" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/selidor-projects/selidor-libraries/selidor-typed-value" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/selidor-projects/selidor-libraries/selidor-typed-value-mapstruct" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/selidor-projects/selidor-parent" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/selidor-projects/selidor-starters" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/selidor-projects/selidor-starters/selidor-starter-parent" charset="UTF-8" />
