+++ selidor-projects/selidor-dependencies/pom.xml
@@ -95,6 +95,11 @@
       <!-- Selidor -->
       <dependency>
         <groupId>pw.itr0.selidor</groupId>
+        <artifactId>selidor-core</artifactId>
+        <version>${revision}</version>
+      </dependency>
+      <dependency>
+        <groupId>pw.itr0.selidor</groupId>
         <artifactId>selidor-identifier</artifactId>
         <version>${revision}</version>
       </dependency>
