+++ selidor-projects/selidor-libraries/selidor-core/src/main/java/pw/itr0/selidor/util/ByteArrayUtil.java
@@ -1,4 +1,4 @@
-package pw.itr0.selidor.identifier.util;
+package pw.itr0.selidor.util;
 
 public class ByteArrayUtil {
